+++ protected/humhub/docs/CHANGELOG.md
@@ -45,6 +45,7 @@ HumHub Change Log
 - Fix: Reset modal dialog size + add `size` option
 - Enh: Added `size` option `ui.modal.Modal.set()`
 - Enh: Use `ContentActiveRecord::getUrl()` for content perma links (if given)
+- Enh: Added `ContentTag` concept for creating content categories/filter on module level
 
 1.2.1 (June 17, 2017)
 - Fix: Invite error in french language